small updates and comments
