+++ vault/src/contract.rs
@@ -1,14 +1,13 @@
-use core::ops::AddAssign;
-
 use crate::{
+    balance::mint_shares,
     flash_loan,
-    math::{
-        compute_deposit_ratio, compute_fee_amount, compute_fee_earned, compute_fee_per_share,
-        compute_shares_amount,
-    },
+    math::{compute_deposit_ratio, compute_fee_amount, compute_shares_amount},
     rewards::{update_fee_per_share_universal, update_rewards},
     storage::*,
-    types::{BatchObj, Error},
+    token_utility::{
+        get_token_client, read_flash_loan_balance, transfer, transfer_into_flash_loan,
+    },
+    types::Error,
 };
 use soroban_sdk::{contractimpl, log, Address, BytesN, Env};
 
@@ -21,7 +20,7 @@ pub trait VaultContractTrait {
         flash_loan_bytes: BytesN<32>,
     ) -> Result<(), Error>;
 
-    fn deposit(e: Env, admin: Address, from: Address, amount: i128) -> Result<i128, Error>;
+    fn deposit(e: Env, admin: Address, from: Address, amount: i128) -> Result<(), Error>;
 
     fn deposit_fees(e: Env, flash_loan: Address, amount: i128) -> Result<(), Error>;
 
@@ -29,10 +28,13 @@ pub trait VaultContractTrait {
 
     fn withdraw_matured(e: Env, admin: Address, addr: Address) -> Result<(), Error>;
 
+    // needs to be re-implemented
     fn get_shares(e: Env, id: Address) -> i128;
 
+    // should be removed by the end of the update
     fn get_increment(e: Env, id: Address) -> Result<i128, Error>;
 
+    // needs to be re-implemented
     fn withdraw(e: Env, admin: Address, to: Address) -> Result<(), Error>;
 }
 
@@ -71,6 +73,7 @@ impl VaultContractTrait for VaultContract {
         let client = get_token_client(&e);
         client.transfer(&flash_loan, &e.current_contract_address(), &amount);
 
+        // update the universal fee per share amount here to avoid the need for a collected_last_recorded storage slot.
         update_fee_per_share_universal(&e, amount);
 
         // add the new fees to the collected slot
@@ -81,8 +84,11 @@ impl VaultContractTrait for VaultContract {
         Ok(())
     }
 
-    fn deposit(e: Env, admin: Address, from: Address, amount: i128) -> Result<i128, Error> {
+    fn deposit(e: Env, admin: Address, from: Address, amount: i128) -> Result<(), Error> {
+        // authenticate the admin and check authorization
         auth_admin(&e, admin)?;
+
+        // we update the rewards before the deposit to avoid the abuse of the collected fees by withdrawing them with liquidity that didn't contribute to their generation.
         update_rewards(&e, from.clone())?;
 
         // construct the token client we'll use later on
@@ -96,27 +102,32 @@ impl VaultContractTrait for VaultContract {
             compute_shares_amount(
                 amount,
                 total_supply,
-                read_flash_loan_balance(&e, &token_client),
+                read_flash_loan_balance(&e, &token_client), // shares are calculated for the liquidity, since fees aren't re-invested as liquidity we only count the flash loan contract's balance.
             )
         };
 
         // transfer the funds into the flash loan
         transfer_into_flash_loan(&e, &token_client, &from, &amount);
 
-        log!(&e, "minting {}", shares);
+        // mint the new shares to the lender
         mint_shares(&e, from, shares);
 
-        Ok(0)
+        Ok(())
     }
 
     fn withdraw_matured(e: Env, admin: Address, addr: Address) -> Result<(), Error> {
+        // authenticate the admin and check authorization
         auth_admin(&e, admin)?;
 
         let token_client = get_token_client(&e);
+
+        // collect all the fees matured by the lender `addr`
         let matured = read_matured_fees_particular(&e, addr.clone());
 
+        // transfer the matured yield to `addr` and update the particular matured fees storage slot
         transfer(&e, &token_client, &addr, matured);
         write_matured_fees_particular(&e, addr, 0);
+
         Ok(())
     }
 
@@ -124,36 +135,6 @@ impl VaultContractTrait for VaultContract {
         auth_admin(&e, admin)?;
         update_rewards(&e, id)?;
 
-        //        transfer(&e, &token_client, &id, lender_fees);
-
-        /*        if let Some(Ok(batch)) = get_batch(&e, to.clone(), batch_n) {
-                    let total_supply = get_tot_supply(&e);
-                    let total_balance = get_token_balance(&e, &token_client);
-
-                    if batch.curr_s < shares {
-                        return Err(Error::InvalidShareBalance);
-                    }
-
-                    let new_deposit = compute_deposit_ratio(batch.deposit, shares, batch.init_s);
-                    let fee_amount = compute_fee_amount(new_deposit, shares, total_supply, total_balance);
-
-                    transfer(&e, &to, fee_amount);
-                    burn_shares(&e, to.clone(), shares, batch_n);
-
-                    let new_tot_supply = total_supply - shares;
-                    let new_tot_bal = total_balance - fee_amount;
-
-                    let new_shares = if total_balance != new_deposit {
-                        compute_shares_amount(new_deposit, new_tot_supply, new_tot_bal - new_deposit)
-                    } else {
-                        compute_shares_amount(new_deposit, total_supply, new_deposit)
-                    };
-
-                    mint_shares(&e, to, new_shares, new_deposit);
-                } else {
-                    return Err(Error::BatchDoesntExist);
-                }
-        */
         Ok(())
     }
 
@@ -163,7 +144,7 @@ impl VaultContractTrait for VaultContract {
         // construct the token client we'll use later on
         let token_client = get_token_client(&e);
 
-        let increment = get_increment(&e, to.clone());
+        /*        let increment = get_increment(&e, to.clone());
 
         let mut amount: i128 = 0;
         let mut temp_supply: i128 = get_tot_supply(&e);
@@ -196,7 +177,7 @@ impl VaultContractTrait for VaultContract {
         let flash_loan_id_bytes = get_flash_loan_bytes(&e);
         let flash_loan_client = flash_loan::Client::new(&e, &flash_loan_id_bytes);
         flash_loan_client.withdraw(&get_contract_addr(&e), &initial_deposit, &to);
-        transfer(&e, &token_client, &to, amount);
+        transfer(&e, &token_client, &to, amount);*/
 
         Ok(())
     }
@@ -204,10 +185,11 @@ impl VaultContractTrait for VaultContract {
     fn get_shares(e: Env, id: Address) -> i128 {
         //        get_collected_last_recorded(&e)
         //read_matured_fees_particular(&e, id)
-        read_fee_per_share_particular(&e, id)
+        //        read_fee_per_share_particular(&e, id)
+        0
     }
 
     fn get_increment(e: Env, id: Address) -> Result<i128, Error> {
-        Ok(get_increment(&e, id))
+        Ok(0)
     }
 }
