+++ vault/src/storage.rs
@@ -4,14 +4,14 @@ use soroban_sdk::{unwrap::UnwrapOptimized, vec, Address, BytesN, ConversionError
 
 use crate::{
     token,
-    types::{BatchKey, BatchObj, DataKey, Error},
+    types::{DataKey, Error},
 };
 
 pub fn get_contract_addr(e: &Env) -> Address {
     e.current_contract_address()
 }
 
-fn put_tot_supply(e: &Env, supply: i128) {
+pub fn put_tot_supply(e: &Env, supply: i128) {
     let key = DataKey::TotSupply;
     e.storage().set(&key, &supply);
 }
@@ -50,16 +50,16 @@ pub fn read_matured_fees_particular(e: &Env, addr: Address) -> i128 {
     let key = DataKey::MaturedFeesParticular(addr);
     e.storage().get(&key).unwrap_or(Ok(0)).unwrap()
 }
-
+/*
+// these two shouldn't be needed
 pub fn put_collected_last_recorded(e: &Env, last_recorded: i128) {
     let key = DataKey::CollectedLastRecorded;
     e.storage().set(&key, &last_recorded);
 }
-
 pub fn get_collected_last_recorded(e: &Env) -> i128 {
     let key = DataKey::CollectedLastRecorded;
     e.storage().get(&key).unwrap_or(Ok(0)).unwrap()
-}
+}*/
 
 pub fn put_fee_per_share_universal(e: &Env, last_recorded: i128) {
     let key = DataKey::FeePerShareUniversal;
@@ -101,16 +101,9 @@ pub fn get_flash_loan_bytes(e: &Env) -> BytesN<32> {
     e.storage().get(&key).unwrap().unwrap()
 }
 
+// should be deprecated
 pub fn get_token_balance(e: &Env, client: &token::Client) -> i128 {
     client.balance(&get_contract_addr(e)) + client.balance(&get_flash_loan(e))
-}
-
-pub fn read_flash_loan_balance(e: &Env, client: &token::Client) -> i128 {
-    client.balance(&get_flash_loan(e))
-}
-
-pub fn transfer(e: &Env, client: &token::Client, to: &Address, amount: i128) {
-    client.transfer(&get_contract_addr(e), to, &amount);
 }
 
 pub fn _transfer_in_vault(e: &Env, from: &Address, amount: &i128) {
@@ -135,30 +128,7 @@ pub fn write_administrator(e: &Env, id: Address) {
     e.storage().set(&key, &id);
 }
 
-pub fn mint_shares(e: &Env, to: Address, shares: i128) {
-    let tot_supply = get_tot_supply(e);
-    put_tot_supply(e, tot_supply + shares);
-
-    let mut balance = read_balance(e, to.clone());
-    balance.add_assign(shares);
-    write_balance(e, to, balance);
-}
-
-pub fn burn_shares(e: &Env, to: Address, shares: i128, batch_n: i128) {
-    let tot_supply = get_tot_supply(e);
-    let key = DataKey::Batch(BatchKey(to, batch_n));
-
-    let mut batch: BatchObj = e.storage().get(&key).unwrap().unwrap();
-    batch.curr_s -= shares;
-    put_tot_supply(e, tot_supply - shares);
-
-    if batch.curr_s == 0 {
-        e.storage().remove(&key); // if there are 0 shares remove the batch
-    } else {
-        e.storage().set(&key, &batch);
-    }
-}
-
+/*
 pub fn put_increment(e: &Env, id: Address, n: i128) {
     e.storage().set(&DataKey::Increment(id), &n);
 }
@@ -168,7 +138,7 @@ pub fn get_increment(e: &Env, id: Address) -> i128 {
         .get(&DataKey::Increment(id))
         .unwrap_or(Ok(0))
         .unwrap()
-}
+}*/
 
 pub fn auth_admin(e: &Env, admin: Address) -> Result<(), Error> {
     if read_admin(e) != admin {
@@ -178,6 +148,7 @@ pub fn auth_admin(e: &Env, admin: Address) -> Result<(), Error> {
     Ok(())
 }
 
+/*
 pub fn get_batch(e: &Env, id: Address, batch_n: i128) -> Option<Result<BatchObj, ConversionError>> {
     let key = DataKey::Batch(BatchKey(id, batch_n));
     e.storage().get(&key)
@@ -189,14 +160,4 @@ pub fn get_initial_deposit(e: &Env, id: Address) -> i128 {
 
 pub fn set_initial_deposit(e: &Env, id: Address, amount: i128) {
     e.storage().set(&DataKey::InitialDep(id), &amount)
-}
-
-pub fn transfer_into_flash_loan(e: &Env, client: &token::Client, from: &Address, amount: &i128) {
-    client.transfer(from, &get_flash_loan(e), amount);
-}
-
-pub fn get_token_client(e: &Env) -> token::Client {
-    token::Client::new(e, &get_token_id(e))
-}
-
-// NEW REWARDING SYSTEM
+}*/
