+++ vault/src/types.rs
@@ -9,13 +9,13 @@ pub enum DataKey {
     FlashLoan,
     FlashLoanB,
     Balance(Address),
-    CollectedLastRecorded,
+    //  CollectedLastRecorded, // deprecated, should be removed
     FeePerShareUniversal,
     FeePerShareParticular(Address),
     MaturedFeesParticular(Address),
-    InitialDep(Address),
-    Batch(BatchKey),
-    Increment(Address),
+    //InitialDep(Address), // deprecated, should be removed
+    //    Batch(BatchKey),     // deprecated, should be removed
+    //    Increment(Address), // deprecated, should be removed
 }
 
 #[contracterror]
@@ -24,18 +24,19 @@ pub enum DataKey {
 pub enum Error {
     VaultAlreadyInitialized = 0,
     InvalidAdminAuth = 1,
-    InvalidShareBalance = 2,
-    BatchDoesntExist = 3,
+    InvalidShareBalance = 2, // needs change
+    BatchDoesntExist = 3,    // depreacted
 }
+/*
+#[derive(Clone)]
+#[contracttype]
+pub struct BatchKey(pub Address, pub i128); // depreacated
 
 #[derive(Clone)]
 #[contracttype]
-pub struct BatchKey(pub Address, pub i128);
-
-#[derive(Clone)]
-#[contracttype]
+// deprecated
 pub struct BatchObj {
     pub init_s: i128,
     pub deposit: i128,
     pub curr_s: i128,
-}
+}*/
